// angular-client/ng-bonk/src/app/model/response/paged-response.model.ts
export interface PagedResponse<T> {
  _embedded: {
    [key: string]: T[];
  };
  _links: {
    self: { href: string };
    next?: { href: string };
    prev?: { href: string };
    first?: { href: string };
    last?: { href: string };
  };
  page: {
    size: number;
    totalElements: number;
    totalPages: number;
    number: number;
  };
}